Add unit tests for interface-utils.ts
Also, some moving around of functions and whitespace changes.
